sending and receiving buffer to/from geometry

+++ src/constants.h
@@ -1,7 +1,7 @@
 #ifndef CONSTANTS_H
 #define CONSTANTS_H
 
-const double EPSILON = 1e-5;
+const float EPSILON = 1e-5;
 const int COPLANAR = 0;
 const int FRONT = 1;
 const int BACK = 2;
