+++ src/vector2.h
@@ -5,14 +5,14 @@ class Vector2
 {
     friend class ThreeBSP;
 public:
-    Vector2(double x=0, double y=0);
+    Vector2(float x=0, float y=0);
     Vector2 *clone() const;
     Vector2 *add(const Vector2 *other);
     Vector2 *sub(const Vector2 *other);
-    Vector2 *multiplyScalar(double t);
+    Vector2 *multiplyScalar(float t);
 
 protected:
-    double x;
-    double y;
+    float x;
+    float y;
 };
 #endif // VECTOR2_H
